changes of bill

+++ POS.DTO/CustomerMappingDTO.cs
@@ -10,7 +10,7 @@ namespace POS.DTO
         public System.Int32 MappingId { get; set; }
         public System.Int32 ItemId { get; set; }
         public System.Int32 CustomerId { get; set; }
-        public System.Decimal Rate { get; set; }
+        public double Rate { get; set; }
         public System.String ItemName { get; set; }
         public System.String CustomerName { get; set; }
     }
